requirements.txt
aiohttp==3.6.3
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
braceexpand==0.1.6
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
discord.py==1.5.0
easydict==1.9
Flask==1.1.2
gevent==20.9.0
greenlet==0.4.17
humanize==3.0.1
idna==2.10
import-expression==1.1.4
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jishaku==1.19.2.202
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.8.0
mccabe==0.6.1
motor==2.3.0
multidict==5.0.0
pycparser==2.20
pylint==2.6.0
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.14.0
requirements-detector==0.7
six==1.15.0
toml==0.10.2
tornado==6.0.4
umongo==3.0.0b10
vermin==0.10.5
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.2
zope.event==4.5.0
zope.interface==5.1.2
